- disable teamtpye patches causing crash - fix wrong condition used - fixup some compiler warnings - fixup some verses checkings - add PR #2231

+++ src/Ext/Aircraft/Body.cpp
@@ -1103,7 +1103,7 @@ int FakeAircraftClass::_Mission_Move()
 	default:
 	{
 		auto Current_Mission_Control = this->GetCurrentMissionControl();
-		return  (Current_Mission_Control->Rate * TICKS_PER_MINUTE) + ScenarioClass::Instance->Random.RandomRanged(0, 2);
+		return  int(Current_Mission_Control->Rate * TICKS_PER_MINUTE) + ScenarioClass::Instance->Random.RandomRanged(0, 2);
 	}
 	case 1:
 	{
